Simplify main layout styles
We were using a "push" div in order to force the footer to the bottom,
and were using a wrapper with a minimum height and negative margins.

The same thing can be accomplished using flex and giving making the
wrapper fill the empty space, which in my humble opinion simplifies the
code and makes it easier to follow.

We could further simplify the code by removing the wrapper div of the
footer wrapper, although I'm not sure the benefits overcome potential
inconveniences caused to other institutions who might have custom styles
based on the existence of these wrappers.